[TAO-2229][Fix] Drop the spec generator; the template is the only source of defaults
Matches tao-analyze-detection-kpi and tao-generate-image-embeddings, which
removed theirs for the same reason. `prepare_unique_neighbor_matching_spec.py`
loaded the template and then overwrote its values from argparse defaults
unconditionally, so `--template` was inert for those fields.

As with the embeddings skill, the seven defaults happened to agree with the
template — global, euclidean, 5, three Nones and an empty rare_class_list — so
nothing was producing a wrong value here. It is the same two-sources-of-truth
shape, and the skill had started documenting the copy-the-template path while
still shipping a generator that could override it.

Checked before deleting: the template already carries every field the generator
wrote, including source/target detection files, detection_format and
exclude_path, so nothing is lost.

`load_yaml` and `validate_config` move into
verify_unique_neighbor_matching_spec.py, which imported them from the deleted
module. Verify now also reports the budget, policy and metric — the mined
parquet is a list of filepaths and records nothing about why those files were
chosen.
